Initial commit: cvnnote-sharp, canvon notes in C#/.NET
This is an alternative take on writing software to use with my
canvon notes file syntax. The previous one was written in Haskell
and simply called cvnnote, but apart from an "elaborate" parser
it only got a demo command-line interface. cvnnote-sharp is planned
to feature a CLI, a GUI and possibly a TUI (text user interface,
think ncurses) as well.

All these files generated by MonoDevelop 5.10.0.871-2 (Debian)
template for a library project.