[GUI] MainWindow: Re-wire close action to CloseFile()
This would appear to be saner than what I initially planned.
(To have multiple windows and use close to close the window.)

We now see that, during the 3s artificial wait, the status bar
does not get updated... Only after that does it get updated.
This is a problem, for load information on slowly loading files
later.