[GUI] MainWindow: Move full file path from title to text field
Having the full path in the window title is not so good, as at least my
window manager shortens it in the window list to just the start path,
although rather the file name (without path) is interesting. (?)

So have only the file name in the window title, and add a Gtk.Entry
that looks a bit like the address bar in a web browser.

Later this "address bar" may become editable and a notes file selected
by directly entering its path, or copy-pasting it from somewhere else.