[GUI] MainWindow: Add a 'Recent' menu, register opened files as recently used
This is automatically synchronized via Gtk.RecentManager

Finding the 'Recent' menu for updating it was a problem, though. The
stetic GUI Designer did not seem to create a predefined variable for
the menu, and iterating through the children and sub-children of the
this.menubar1 was problematic due to the casts, method access,
further cast etc. involved, so there is much debugging code still
lying around there. It could be removed / trimmed down, later.