[lib, lib tests] Actually parse day intro (into date, chapter, comment)
It's nice that the design seems to be modular enough this time
that no changes to the GUI were necessary, while the GUI automatically
reflects the changes made to the lib. The detailed test cases even
compiled, although they needed slight adjustments to succeed again.

This now finds some small amount of problems in a real-world cvnnote
file with roughly one month's worth of notes, but mostly seems to work.
The parsing will have to be improved slightly, to cope with "[...]"
pseudo-days, and an uninitialized variable bug for the top-level Notes
object needs to be fixed. A "chapter"/comment of "(So)" (Sonntag,
Sunday) will need a bit of thinking whether I should accept this as
correct at all and how to parse it. Also I wanted the CvnNote suite
to disclose week day names anyhow, but this didn't work yet, at least
with my (mixed en_GB <-> de_DE) locale.