[licensing] Be bold and claim copyright on AssemblyInfo.cs files
.., to make sure my intent to license my work under GPL3 is clear,
including the meta-data I entered for the produced assemblies.

I also hope that, under substantial changes, this meta-data would
need to be changed per the license, so the changed version would
not be confused to be endorsed by myself. I hope for this to happen
I would not have needed to choose another license, like zlib license
or something.